Building WebRTC native library for our project
Compiling WebRTC native library to get the library and its interface
classes then add those things to the project:

1.Follow the instructions exactly https://webrtc.org/native-code/ios/
(This might take atleast 1 day because this downloads source code of
size 13 GB the we have to compile it)
2.After following all the steps proceed as below
i.Drag libwebrtc.a found at src/out_ios64/Debug-iphoneos to the
project.This is library basically that has webRTC functions.
ii.Drag all the files found at
/src/chromium/src/third_party/libjingle/source/talk/app/webrtc/objc/publ
ic to the project.These are the source that interacts with the above
library.